Created report form
Created the report form and linked it to the Report on merge menu item.
Still need to set up the ReportButton event.  Currently reports aren't
saved or sent anywhere, that will be taken care of later.

I still want to incorporate something so users can see the current
reported rating for the plugin they're submitting a report for.